Fold schema/catalog inspection into env-gen runner
Merge try_environment_intent_schema.py into environment_generation_runner.py:
add 'schema' and 'catalog' modes that short-circuit before Isaac Sim, and
always print the tabular graph view after resolving. Consolidate the agent
call into resolve_env_spec so the prompt-to-spec path lives in one function.
