+++ isaaclab_arena_examples/agentic_environment_generation/environment_generation_runner.py
@@ -7,6 +7,12 @@
 
 Usage::
 
+    # Print the Pydantic ArenaEnvGraphSpec JSON schema (no agent call, no Isaac Sim):
+    python isaaclab_arena_examples/agentic_environment_generation/environment_generation_runner.py --mode schema
+
+    # Print the catalog sent to the agent (no agent call, no Isaac Sim):
+    python isaaclab_arena_examples/agentic_environment_generation/environment_generation_runner.py --mode catalog
+
     # Resolve a prompt into an environment graph spec YAML (no Isaac Sim):
     python isaaclab_arena_examples/agentic_environment_generation/environment_generation_runner.py --mode resolve --prompt ...
 
@@ -47,11 +53,13 @@ def add_agentic_env_gen_runner_cli_args(parser: argparse.ArgumentParser) -> None
     group.add_argument(
         "--mode",
         type=str,
-        choices=("full", "resolve", "build"),
+        choices=("full", "resolve", "build", "schema", "catalog"),
         default="full",
         help=(
-            "Which phases to run: 'resolve' (no Isaac Sim), 'build' (needs --env_graph_spec_yaml), "
-            "or 'full' (resolve and build in one process; default)."
+            "Which phases to run: 'schema' (print the spec JSON schema and exit), "
+            "'catalog' (print the agent catalog and exit), 'resolve' (prompt -> spec YAML, no Isaac Sim), "
+            "'build' (needs --env_graph_spec_yaml), or 'full' (resolve and build in one process; default). "
+            "'schema' and 'catalog' make no agent call."
         ),
     )
     group.add_argument(
@@ -86,8 +94,8 @@ def add_agentic_env_gen_runner_cli_args(parser: argparse.ArgumentParser) -> None
     )
 
 
-def generate_env_graph_spec(args_cli: argparse.Namespace) -> ArenaEnvGraphSpec:
-    """Generate an environment graph spec from a prompt."""
+def resolve_env_spec(args_cli: argparse.Namespace) -> Path:
+    """Resolve a prompt into an environment graph spec YAML."""
     from isaaclab_arena.agentic_environment_generation.environment_generation_agent import (
         EnvironmentGenerationAgent,
         build_asset_catalogue,
@@ -120,19 +128,69 @@ def generate_env_graph_spec(args_cli: argparse.Namespace) -> ArenaEnvGraphSpec:
         invalid_path = write_env_graph_dict(data, args_cli.out_dir)
         print(f"[runner] wrote invalid spec YAML to {invalid_path}", flush=True)
         assert False, f"Agent returned an invalid spec. Validation traces: {agent.last_validation_traces}"
+    print_env_graph(env_graph_spec)
     print(
         f"[runner] generated → {env_graph_spec.summary()}, env_name={env_graph_spec.env_name!r}",
         flush=True,
     )
-    return env_graph_spec
-
-
-def resolve_env_spec(args_cli: argparse.Namespace) -> Path:
-    """Resolve a prompt into an environment graph spec YAML."""
-    env_graph_spec = generate_env_graph_spec(args_cli)
     path = write_env_graph_spec(env_graph_spec, args_cli.out_dir)
     print(f"[runner] wrote environment graph spec → {path}", flush=True)
     return path
+
+
+def print_schema() -> None:
+    """Print the Pydantic ArenaEnvGraphSpec JSON schema."""
+    import json
+
+    from isaaclab_arena.environments.arena_env_graph_spec import ArenaEnvGraphSpec
+
+    print(json.dumps(ArenaEnvGraphSpec.model_json_schema(), indent=2))
+
+
+def print_catalog() -> None:
+    """Print the asset, relation, and task catalogs sent to the agent."""
+    from isaaclab_arena.agentic_environment_generation.environment_generation_agent import (
+        build_asset_catalogue,
+        build_relation_catalogue,
+        build_task_catalogue,
+    )
+
+    print(build_asset_catalogue().to_catalog_string())
+    print()
+    print(build_relation_catalogue().to_catalog_string())
+    print()
+    print(build_task_catalogue().to_catalog_string())
+
+
+def _iter_printable_assets(spec: ArenaEnvGraphSpec):
+    yield "embodiment", spec.embodiment.id, spec.embodiment.registry_name, spec.embodiment.params
+    yield "background", spec.background.id, spec.background.registry_name, spec.background.params
+    for obj in spec.objects:
+        yield "object", obj.id, obj.registry_name, obj.params
+    for ref in spec.object_references or []:
+        yield "object_reference", ref.id, ref.id, ref.params
+
+
+def print_env_graph(spec: ArenaEnvGraphSpec) -> None:
+    """Print the generated graph in a human-readable tabular layout."""
+    print(f"\n=== ArenaEnvGraphSpec (env_name={spec.env_name!r}) ===")
+
+    print("\nassets:")
+    for role, asset_id, registry_name, params in _iter_printable_assets(spec):
+        params_str = f"  params={params}" if params else ""
+        print(f"  {asset_id:24s} role={role:18s} registry_name={registry_name}{params_str}")
+
+    print("\nrelations:")
+    for relation in spec.relations:
+        ref_str = f"  reference={relation.reference}" if relation.reference is not None else ""
+        params_str = f"  params={relation.params}" if relation.params else ""
+        print(f"  {relation.kind:16s} subject={relation.subject}{ref_str}{params_str}")
+
+    print(f"\ntask: composition={spec.task.composition}")
+    print(f"  description: {spec.task.description}")
+    for i, task in enumerate(spec.task.subtasks):
+        print(f"  [{i}] kind={task.kind}")
+        print(f"    params: {task.params}")
 
 
 def build_env_from_env_graph_spec(env_graph_spec_path: Path, args_cli: argparse.Namespace) -> ManagerBasedEnv:
@@ -192,6 +250,14 @@ def main() -> int:
     parser = get_isaaclab_arena_cli_parser()
     add_agentic_env_gen_runner_cli_args(parser)
     args_cli = parser.parse_args()
+
+    if args_cli.mode == "schema":
+        print_schema()
+        return 0
+
+    if args_cli.mode == "catalog":
+        print_catalog()
+        return 0
 
     if args_cli.mode == "resolve":
         resolve_env_spec(args_cli)
